removed modelTemplates.ts, moved primitives to oglPrimitives.ts and createRandomObjectDescription to objectDescription.ts
(cherry picked from commit bc8d1c9765d820f5819e29ec2b1b16a941ef4780)
