+++ src/core/engines/createRenderingEngine.ts
@@ -1,8 +1,8 @@
 /*
-  (c) 2021 Open AR Cloud
+  (c) 2026 Open AR Cloud
   This code is licensed under MIT license (see LICENSE.md for details)
 
-  (c) 2024 Nokia
+  (c) 2026 Nokia
   Licensed under the MIT License
   SPDX-License-Identifier: MIT
 */
